// frontend/src/shared/api/generated/model/examOut.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * novalearn-api
 * OpenAPI spec version: 0.1.0
 */
import type { RoadmapNodeOut } from './roadmapNodeOut';

export interface ExamOut {
  id: string;
  user_id: string | null;
  folder_id: string;
  name: string;
  exam_date: string;
  roadmap_nodes?: RoadmapNodeOut[];
  created_at: string;
  progress?: number;
}